Return Sea Level plot

Return Level Plot for Sea Level and convolution.

Usage

RSLplot(data,
           z = NULL,
           duration = NULL,
           lambda,
           xlab = "period", ylab = "level",
           mono  = TRUE,
           lty.quant = "solid",
           col.quant = ifelse(mono, "black", "SteelBlue"),
           pct.conf = c(95, 70),
           col.conf = NULL,
           filled.conf = FALSE,
           fill.conf = NULL,
           lty.conf = ifelse(rep(filled.conf, 2),
                             rep(NA, 2), c("dashed", "dotted")),
           rl.mark = 100,
           text.mark = rl.mark,
           col.mark = NULL,
           Tlim = NULL,
           problim = NULL,
           below = NA,
           alpha.below = 0.5,
           grid = TRUE,
           legend = TRUE,
           pch.points = c(24, 22, 23),
           col.points = c("black", "darkgray"),
           ...)

Arguments

data

Data frame containing the description of the return level curve and confidence limits. See Details.

z

Empirical points to add to the plot.

duration

For empirical points.

lambda

Event rate (in events by year). This is used only when empirical points are added to the plot using the x argument.

xlab

Label for the x-axis.

ylab

Label for the y-axis.

mono

Monochrome plot.

lty.quant

Line type for the quantile (return level curve).

col.quant

Line color for the quantile (return level curve).

pct.conf

Confidence levels (in accordance with the content of the data.frame data.

col.conf

Line colors for the confidence limits, in the same order as pct.conf.

filled.conf

Should confidence regions be filled?

fill.conf

Colors for the confidence regions if filled.

lty.conf

Line types for confidence regions limits. By using NA, the plot of the limits (lower and upper) will be cancelled which may be when fill.conf is TRUE.

rl.mark

Return level(s) to materialize with a vertical line.

text.mark

Label(s) to display near the mark.rl lines if any.

col.mark

Color(s) for the marK.rl.

Tlim

Limits for the x-axis given in time (return periods).

problim

Limits for the x-axis given in probability. Only one of the two arguments Tlim and problim can be given.

below

A level below which a semi-transparent rectangle will be drawn, with the meaning that the corresponding return levels should not be used. In the sea-level convolution context, this will normally be HAT + threshold because the distribution of the sea-level is conditional to a surge level exceeding the threshold.

alpha.below

A value of transparency to (partially) occult the region of the plot lying below the minimal level of validity for the convolution, see Details. The value 0 means a fully transparent rectangle (no effect), and the value 1 means a fully opaque one. When using a device that do not support transparency, only the values 0 and 1 will be possible.

grid

Logical. If TRUE, a grid with dotted lines is added.

legend

Logical. If TRUE, a legend is added.

pch.points

Vector of symbols to be used for points given by z (if any). When z is a numeric vector, only the first element is used. When z is a list, the given symbols will be used in correspondence with the list elements.

col.points

Vector of colours used in the same way as pch.points.

...

Other arguments to pass to plot. These can be main, ylim, etc.

Details

This function is roughly a copy of the RLplot in the package Renext with a few supplementary formal arguments.

Line types and colors for confidence regions are given as a vectors with elements corresponding to the percents given in pct.conf. These levels should be available in the data.frame data, which means that suitable columns should indicate the limits. For instance a 95 percent region can be displayed only when the two columns L.95 and U.95 exist and provide the relevant information.

Value

None.

Author

Yves Deville

See also